move contacts transform into constants
